requirements.txt
selenium==4.18.1
bs4==0.0.2
openai==1.13.3
webdriver-manager==4.0.1
pipenv==2023.12.1
#
python-dotenv==1.0.1